requirements.txt
asgiref==3.8.1
Django==4.2.13
djangorestframework==3.15.1
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.11.0
gunicorn==22.0.0